Add files via upload
RSVP Movies, a prominent Indian film production company renowned for its super-hit movies, is embarking on a new venture in 2022 by creating a film tailored for a global audience. To ensure the success of this project, they have turned to data analytics. As a data analyst and SQL expert, I have been entrusted with the task of analyzing the movie data from the past three years and providing valuable insights to guide RSVP Movies in their strategic decisions.